Switch latency_stddev to be samplerate independent (and thus in ms).
Visualization seems correct but I didn't verify the math.

+++ src/latencyfilter.h
@@ -44,6 +44,6 @@ public:
 private:
 	Settings& settings;
 	Random& random;
-	double latency_offset{0.0};
+	float latency_offset{0.0};
 	std::size_t latency_last_pos{0};
 };
